register and login

+++ routers/todos.py
@@ -7,6 +7,7 @@ from fastapi import (
     HTTPException,
     status,
     Request,
+    Form,
 )
 from sqlalchemy.orm import Session
 from typing import Annotated
@@ -47,21 +48,48 @@ class TodoRequest(BaseModel):
 @router.get("/", response_class=HTMLResponse)
 async def read_all_by_user(request: Request, user: user_dependency, db: db_dependency):
     if user is None:
-        raise HTTPException(
-            status_code=401, detail="Invalid authentication credentials"
-        )
+        return RedirectResponse(url="/auth", status_code=status.HTTP_302_FOUND)
     todos = (
         db.query(Todos)
         .filter(Todos.owner_id == user.get("user_id"))
         .order_by(Todos.priority.desc())
         .all()
     )
-    return templates.TemplateResponse("home.html", {"request": request, "todos": todos})
+    return templates.TemplateResponse(
+        "home.html", {"request": request, "user": user, "todos": todos}
+    )
 
 
 @router.get("/add-todo", response_class=HTMLResponse)
 async def add_new_todo(request: Request):
-    return templates.TemplateResponse("add-todo.html", {"request": request})
+    user = await get_current_user(request)
+    if user is None:
+        return RedirectResponse(url="/auth", status_code=status.HTTP_302_FOUND)
+    return templates.TemplateResponse(
+        "add-todo.html", {"request": request, "user": user}
+    )
+
+
+@router.post("/add-todo", response_class=HTMLResponse)
+async def create_new_todo(
+    request: Request,
+    db: db_dependency,
+    title: str = Form(...),
+    description: str = Form(...),
+    priority: str = Form(...),
+):
+    user = await get_current_user(request)
+    if user is None:
+        return RedirectResponse(url="/auth", status_code=status.HTTP_302_FOUND)
+    todo = Todos(
+        title=title,
+        description=description,
+        priority=priority,
+        owner_id=user.get("user_id"),
+    )
+    db.add(todo)
+    db.commit()
+    return RedirectResponse(url="/todos", status_code=status.HTTP_302_FOUND)
 
 
 @router.get("/edit-todo/{todo_id}", response_class=HTMLResponse)
@@ -72,9 +100,7 @@ async def edit_todo(
     todo_id: int = Path(gt=0),
 ):
     if user is None:
-        raise HTTPException(
-            status_code=401, detail="Invalid authentication credentials"
-        )
+        return RedirectResponse(url="/auth", status_code=status.HTTP_302_FOUND)
     todo_item = (
         db.query(Todos)
         .filter(Todos.id == todo_id)
@@ -83,7 +109,7 @@ async def edit_todo(
     )
 
     return templates.TemplateResponse(
-        "edit-todo.html", {"request": request, "todo": todo_item}
+        "edit-todo.html", {"request": request, "user": user, "todo": todo_item}
     )
 
 
@@ -91,12 +117,15 @@ async def edit_todo(
 async def edit_todo(
     request: Request,
     db: db_dependency,
+    user: user_dependency,
     todo_id: int,
     title: str,
     description: str,
     priority: int,
     completed: bool,
 ):
+    if user is None:
+        return RedirectResponse(url="/auth", status_code=status.HTTP_302_FOUND)
     todo_item = db.query(Todos).filter(Todos.id == todo_id).first()
     todo_item.title = title
     todo_item.description = description
@@ -111,8 +140,12 @@ async def edit_todo(
 async def delete_todo(
     request: Request,
     db: db_dependency,
+    user: user_dependency,
     todo_id: int,
 ):
+    if user is None:
+        return RedirectResponse(url="/auth", status_code=status.HTTP_302_FOUND)
+
     todo_item = db.query(Todos).filter(Todos.id == todo_id).first()
     db.delete(todo_item)
     db.commit()
@@ -120,7 +153,11 @@ async def delete_todo(
 
 
 @router.get("complete/{todo_id}", response_class=HTMLResponse)
-async def complete_todo(request: Request, db: db_dependency, todo_id: int):
+async def complete_todo(
+    request: Request, user: user_dependency, db: db_dependency, todo_id: int
+):
+    if user is None:
+        return RedirectResponse(url="/auth", status_code=status.HTTP_302_FOUND)
     todo_item = db.query(Todos).filter(Todos.id == todo_id).first()
     todo_item.completed = True
     db.add(todo_item)
